added chron job to insert performance score of schemes

+++ backend/analytics-service/src/main/java/org/arghyam/jalsoochak/analytics/dto/event/SchemePerformanceEvent.java
@@ -4,6 +4,8 @@ import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
+import java.math.BigDecimal;
+
 @Data
 @NoArgsConstructor
 @JsonIgnoreProperties(ignoreUnknown = true)
@@ -12,6 +14,6 @@ public class SchemePerformanceEvent {
     private String eventType;
     private Integer tenantId;
     private Integer schemeId;
-    private Integer performanceScore;
+    private BigDecimal performanceScore;
     private String lastWaterSupplyDate;
 }
